Updated the test file structure
